requirements.txt
crewai[tools]
langchain-groq
python-dotenv
google-auth-oauthlib
google-api-python-client
langchain-community
litellm


fastapi
uvicorn
motor
pymongo
authlib
itsdangerous
python-jose[cryptography]
python-multipart